+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## [1.2.5](https://github.com/ThinAirThings/uix/compare/v1.2.4...v1.2.5) (2024-05-23)
 
+## 4.0.0-next.5
+
+### Minor Changes
+
+- add nested error handling for drafts
+
+### Patch Changes
+
+- 13493c5: fix reverse lookup
+
 ## 4.0.0-next.1
 
 ### Minor Changes